+++ src/parse.hpp
@@ -2,7 +2,7 @@
 #include "config.hpp"
 
 int parse_int(std::string value);
-int parse_double(std::string value);
+double parse_double(std::string value);
 
 wf_key parse_key(std::string value);
 wf_button parse_button(std::string value);
